change ints to Long, just to be sure

+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/cart/data/CartItem.java
@@ -17,9 +17,9 @@ public class CartItem {
     private Long id;
     @ManyToOne
     private Product product;
-    private int amount;
+    private Long amount;
 
-    public CartItem(Product product, int amount) {
+    public CartItem(Product product, Long amount) {
         this.product = product;
         this.amount = amount;
     }
