+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/product/web/ProductResponse.java
@@ -7,7 +7,7 @@ public class ProductResponse {
     private final Long id;
     private final String name;
     private final String description;
-    private final int amount;
+    private final Long amount;
     private final String unit;
     private final double price;
